feat: Update User model and fix manager assignment; fix backup task to execute with container name generic local and prod compose
