+++ config/settings/production.py
@@ -1,4 +1,6 @@
 # ruff: noqa: E501
+from celery.schedules import crontab
+
 from .base import *  # noqa: F403
 from .base import DATABASES
 from .base import INSTALLED_APPS
@@ -200,5 +202,12 @@ LOGGING = {
 SPECTACULAR_SETTINGS["SERVERS"] = [
     {"url": "https://example.com", "description": "Production server"},
 ]
-# Your stuff...
-# ------------------------------------------------------------------------------
+
+CELERY_TASK_EAGER_PROPAGATES = False
+
+CELERY_BEAT_SCHEDULE = {
+    "backup_postgres_diario": {
+        "task": "order_service.core.tasks.generate_backup_postgres",
+        "schedule": crontab(hour=2, minute=0),
+    },
+}
