+++ docker-compose.local.yml
@@ -8,8 +8,9 @@ services:
     build:
       context: .
       dockerfile: ./compose/local/django/Dockerfile
+
+    container_name: order_service_django
     image: order_service_local_django
-    container_name: order_service_local_django
     depends_on:
       - postgres
       - redis
@@ -32,8 +33,9 @@ services:
     build:
       context: .
       dockerfile: ./compose/production/postgres/Dockerfile
+
+    container_name: order_service_postgres
     image: order_service_production_postgres
-    container_name: order_service_local_postgres
     volumes:
       - order_service_local_postgres_data:/var/lib/postgresql/data
       - order_service_local_postgres_data_backups:/backups
@@ -50,14 +52,14 @@ services:
 
   redis:
     image: docker.io/redis:6
-    container_name: order_service_local_redis
+    container_name: order_service_redis
     volumes:
       - order_service_local_redis_data:/data
 
   celeryworker:
     <<: *django
     image: order_service_local_celeryworker
-    container_name: order_service_local_celeryworker
+    container_name: order_service_celeryworker
     volumes:
       - /var/run/docker.sock:/var/run/docker.sock
       - order_service_local_postgres_data_backups:/backups
@@ -75,7 +77,7 @@ services:
 
   celerybeat:
     <<: *django
-    image: order_service_local_celerybeat
+    image: order_service_celerybeat
     container_name: order_service_local_celerybeat
     volumes:
       - /var/run/docker.sock:/var/run/docker.sock
