+++ order_service/core/admin.py
@@ -4,11 +4,9 @@ from .models import ComplementaryOrder
 from .models import DeliveryPerson
 from .models import Order
 from .models import OrderTracking
-from .models import User
 
 # Register your models here.
 admin.site.register(Order)
-admin.site.register(User)
 admin.site.register(OrderTracking)
 admin.site.register(DeliveryPerson)
 admin.site.register(ComplementaryOrder)
