+++ order_service/core/api/views.py
@@ -2,19 +2,14 @@ from django.db import transaction
 from rest_framework import viewsets
 
 from order_service.core.models import ComplementaryOrder
-from order_service.core.models import DeliveryPerson
 from order_service.core.models import Establishment
 from order_service.core.models import Order
 from order_service.core.models import OrderTracking
-from order_service.core.models import User
 from order_service.services import order_tracking_service
 
 from .serializers import CreateComplementaryOrderSerializer
 from .serializers import CreatedEstablishmentSerializer
 from .serializers import CreateOrderAlignedSerializer
-from .serializers import DeliveryPersonCreatedSerializer
-from .serializers import DeliveryPersonReadSerializer
-from .serializers import DeliveryPersonUpdateSerializer
 from .serializers import OrderCreatedSerializer
 from .serializers import OrderReadSerializer
 from .serializers import OrderTrackingCreatedSerializer
@@ -25,104 +20,6 @@ from .serializers import ReadEstablishmentSerializer
 from .serializers import ReadOnlyComplementaryOrderSerializer
 from .serializers import UpdateComplementaryOrderSerializer
 from .serializers import UpdateEstablishmentSerializer
-from .serializers import UserCreatedSerializer
-from .serializers import UserReadSerializer
-from .serializers import UserUpdateSerializer
-
-
-class UserViewSet(viewsets.ModelViewSet):
-    """
-    ViewSet for managing User instances.
-
-    Handles standard CRUD operations (Create, Retrieve, Update, Delete) for the User model.
-    Automatically selects the appropriate serializer based on the current action.
-
-    Serializer classes:
-        - UserCreatedSerializer: Used when creating a new user (POST).
-        - UserUpdateSerializer: Used for updating user data (PUT/PATCH).
-        - UserReadSerializer: Default serializer for read operations (GET).
-
-    Query parameter filters:
-        - is_active (bool): Filters users by active status.
-            Example: ?is_active=true
-        - user_type (str): Filters users by role/type.
-            Options: "Customer", "Administrator", "Delivery Man"
-            Example: ?user_type=Customer
-
-    Example endpoints:
-        - GET    /api/users/           → List users
-        - POST   /api/users/           → Create a new user
-        - GET    /api/users/{id}/      → Retrieve a specific user
-        - PUT    /api/users/{id}/      → Fully update a user
-        - PATCH  /api/users/{id}/      → Partially update a user
-        - DELETE /api/users/{id}/      → Delete a user
-    """
-
-    serializer_class = UserReadSerializer
-    queryset = User.objects.all()
-
-    def get_serializer_class(self):
-        if self.action == "create":
-            return UserCreatedSerializer
-        if self.action in ["update", "partial_update"]:
-            return UserUpdateSerializer
-        return UserReadSerializer
-
-    def get_queryset(self):
-        queryset = super().get_queryset()
-
-        is_active = self.request.query_params.get("is_active", None)
-        if is_active is not None:
-            if is_active.lower() == "true":
-                queryset = queryset.active()
-            elif is_active.lower() == "false":
-                queryset = queryset.inactive()
-
-        user_type = self.request.query_params.get("user_type", None)
-        if user_type is not None:
-            if user_type == "Customer":
-                queryset = queryset.customers()
-            elif user_type == "Administrator":
-                queryset = queryset.administrator
-            elif user_type == "Delivery Man":
-                queryset = queryset.delivery_man
-
-        return queryset
-
-
-class DeliveryPersonViewSet(viewsets.ModelViewSet):
-    """
-    ViewSet for managing DeliveryPerson instances.
-
-    Provides full CRUD operations (Create, Retrieve, Update, Delete) for delivery personnel.
-    Automatically selects the appropriate serializer class based on the type of action.
-
-    Serializer classes:
-        - DeliveryPersonCreatedSerializer: Used when creating a new delivery person (POST).
-        - DeliveryPersonUpdateSerializer: Used for updating delivery person data (PUT/PATCH).
-        - DeliveryPersonReadSerializer: Default serializer for read operations (GET).
-
-    Query optimizations:
-        - Uses `select_related("user")` to optimize database access by prefetching related user data.
-
-    Example endpoints:
-        - GET    /api/delivery-people/           → List delivery personnel
-        - POST   /api/delivery-people/           → Create a new delivery person
-        - GET    /api/delivery-people/{id}/      → Retrieve a specific delivery person
-        - PUT    /api/delivery-people/{id}/      → Fully update a delivery person
-        - PATCH  /api/delivery-people/{id}/      → Partially update a delivery person
-        - DELETE /api/delivery-people/{id}/      → Delete a delivery person
-    """
-
-    queryset = DeliveryPerson.objects.all().select_related("user")
-    serializer_class = DeliveryPersonReadSerializer
-
-    def get_serializer_class(self):
-        if self.action == "create":
-            return DeliveryPersonCreatedSerializer
-        if self.action in ["update", "partial_update"]:
-            return DeliveryPersonUpdateSerializer
-        return DeliveryPersonReadSerializer
 
 
 class OrderViewSet(viewsets.ModelViewSet):
