+++ order_service/core/models.py
@@ -1,8 +1,12 @@
 # Create your models here.
+from django.conf import settings
 from django.db import models
 from django.utils import timezone
 
-from . import querysets
+from order_service.users import querysets
+from order_service.users.models import DeliveryPerson
+
+auth_user_model = settings.AUTH_USER_MODEL
 
 
 # Models Abstract
@@ -20,42 +24,6 @@ class TimeStampedModel(CreatedAtModel):
 
     class Meta:
         abstract = True
-
-
-# Models Main
-
-
-class UserType(models.IntegerChoices):
-    ADMINISTRATOR = 1, "Administrator"
-    CUSTOMER = 2, "Customer"
-    DELIVERY_MAN = 3, "Delivery Man"
-    System = 4, "System"  # Validar se seria Interesssante
-
-
-class User(TimeStampedModel):
-    full_name = models.CharField(max_length=250)
-    cpf = models.CharField(max_length=11, unique=True)
-    birth_date = models.DateField()
-    email = models.EmailField(max_length=250)
-    password = models.CharField(max_length=50)
-    is_active = models.BooleanField(default=True)
-
-    user_type = models.IntegerField(choices=UserType)
-
-    objects = querysets.UserQuerySet.as_manager()
-
-    def __str__(self):
-        return self.full_name
-
-
-class DeliveryPerson(TimeStampedModel):
-    user = models.OneToOneField(User, on_delete=models.CASCADE)
-    availability = models.CharField(max_length=50)
-    vehicle = models.CharField(max_length=50)
-    license_plate = models.CharField(max_length=50)
-
-    def __str__(self):
-        return f"Delivery Person {self.user.full_name}"
 
 
 class Address(models.Model):
@@ -133,18 +101,18 @@ class OrderTracking(models.Model):
 
 
 class UserNotification(TimeStampedModel):
-    user = models.ForeignKey(User, on_delete=models.CASCADE)
+    user = models.ForeignKey(auth_user_model, on_delete=models.CASCADE)
     message = models.TextField()
     delivery_channel = models.CharField(max_length=50)
     sent_at = models.DateTimeField(default=timezone.now)
     is_read = models.BooleanField(default=False)
 
     def __str__(self):
-        return f"Notification for {self.user.full_name}"
+        return f"Notification for {self.user.name}"
 
 
 class UserLog(models.Model):
-    user = models.ForeignKey(User, on_delete=models.CASCADE)
+    user = models.ForeignKey(auth_user_model, on_delete=models.CASCADE)
     action = models.CharField(max_length=50)
     msg = models.JSONField(null=True, blank=True)
     timestamp = models.DateTimeField(default=timezone.now)
@@ -154,7 +122,7 @@ class UserLog(models.Model):
 
 
 class AccessLog(models.Model):
-    user = models.ForeignKey(User, on_delete=models.CASCADE)
+    user = models.ForeignKey(auth_user_model, on_delete=models.CASCADE)
     action = models.CharField(max_length=50)
     origin_ip = models.CharField(max_length=45)
     timestamp = models.DateTimeField(default=timezone.now)
